Add markdown format (shell doc)

+++ mk_sh_doc.sh
@@ -1,5 +1,8 @@
 #
-# Extract doc from a shell script, then generate the TOC and save to an HTML file
+# Extract doc from a shell script
+# If generating html, build a TOC
+#
+# Syntax : mk_sh_doc.sh <source> <target> <format (html|md)> [options]
 #
 #============================================================================
 
@@ -10,23 +13,26 @@ SOURCE=$1
 shift
 TARGET=$1
 shift
+FORMAT=$1
+shift
 OPTIONS=$*
 PPC="$PHP $PPC_DIR/ppc.php"
-XTRACT="$PHP $PPC_DIR/extract_shell_doc.php"
+XTRACT="$PHP $PPC_DIR/extract_shell_doc.php -f $FORMAT"
 
-export PHP PPC_DIR SOURCE TARGET OPTIONS PPC XTRACT
+export PHP PPC_DIR SOURCE TARGET OPTIONS PPC XTRACT FORMAT
 
 #-------------
 
 (
-echo '<p>{toc}</p><p>&nbsp;</p><hr/>'
-[ -f src/$TARGET.start.htm ] && cat src/$TARGET.start.htm
+[ "$FORMAT" = html ] && echo '<p>{toc}</p><p>&nbsp;</p><hr/>'
+[ -f src/$TARGET.start.$FORMAT ] && cat src/$TARGET.start.$FORMAT
 $XTRACT $OPTIONS $SOURCE
-[ -f src/$TARGET.end.htm ] && cat src/$TARGET.end.htm
-) >$TARGET.tmp
+[ -f src/$TARGET.end.$FORMAT ] && cat src/$TARGET.end.$FORMAT
+) >$TARGET
 
-PPC_DIR=. $PPC $TARGET.tmp
-mv _$TARGET.tmp $TARGET
-/bin/rm -f $TARGET.tmp
-
-
+if [ "$FORMAT" = html ] ; then
+	mv $TARGET $TARGET.tmp
+	PPC_DIR=. $PPC $TARGET.tmp
+	mv _$TARGET.tmp $TARGET
+	/bin/rm -f $TARGET.tmp
+fi
